+++ globalpayments-android-sdk/src/main/java/com/globalpayments/android/sdk/model/FormValid.kt
@@ -1,0 +1,11 @@
+package com.globalpayments.android.sdk.model
+
+data class FormValid(
+    val numberValid: Boolean = false,
+    val cvvValid: Boolean = false,
+    val dateValid: Boolean = false,
+    val holderValid: Boolean = false,
+) {
+    val cardValid: Boolean
+        get() = numberValid && cvvValid && dateValid && holderValid
+}