+++ globalpayments-android-sdk/src/main/java/com/globalpayments/android/sdk/task/TaskResult.kt
@@ -1,0 +1,6 @@
+package com.globalpayments.android.sdk.task
+
+sealed interface TaskResult<out T> {
+    data class Success<T>(val data: T) : TaskResult<T>
+    data class Error(val exception: Exception) : TaskResult<Nothing>
+}