+++ sample-app/src/main/java/com/globalpayments/android/sdk/sample/gpapi/GPAPIMainFragment.java
@@ -1,8 +1,10 @@
 package com.globalpayments.android.sdk.sample.gpapi;
 
 import android.widget.Button;
+
 import androidx.annotation.IdRes;
 import androidx.fragment.app.Fragment;
+
 import com.globalpayments.android.sdk.sample.R;
 import com.globalpayments.android.sdk.sample.common.base.BaseFragment;
 import com.globalpayments.android.sdk.sample.common.views.CustomToolbar;
@@ -12,6 +14,7 @@ import com.globalpayments.android.sdk.sample.gpapi.batch.closeBatch.CloseFragmen
 import com.globalpayments.android.sdk.sample.gpapi.configuration.GPAPIConfigurationFragment;
 import com.globalpayments.android.sdk.sample.gpapi.deposits.DepositsFragment;
 import com.globalpayments.android.sdk.sample.gpapi.disputes.DisputesFragment;
+import com.globalpayments.android.sdk.sample.gpapi.netcetera.NetceteraFragment;
 import com.globalpayments.android.sdk.sample.gpapi.paymentmethod.PaymentMethodsFragment;
 import com.globalpayments.android.sdk.sample.gpapi.transaction.TransactionFragment;
 import com.globalpayments.android.sdk.sample.gpapi.verifications.VerificationsFragment;
@@ -38,6 +41,7 @@ public class GPAPIMainFragment extends BaseFragment {
         setButtonClickListener(R.id.disputesButton, new DisputesFragment());
         setButtonClickListener(R.id.actionsButton, new ActionsReportFragment());
         setButtonClickListener(R.id.closeBatchButton, new CloseFragment());
+        setButtonClickListener(R.id.hosted_fields_button, new NetceteraFragment());
     }
 
     private void setButtonClickListener(@IdRes int buttonId, Fragment fragment) {